fix: actualizar tipo de dato stock de Integer a Long en DTOs y Entidades

+++ backend/src/main/java/com/stockSync/backend/dto/ProductResponse.java
@@ -14,7 +14,7 @@ public class ProductResponse {
     private String name;
     private String description;
     private BigDecimal price;
-    private Integer stock;
+    private Long stock;
     private String sku;
     private String imageUrl;
     private Boolean active;
@@ -25,7 +25,7 @@ public class ProductResponse {
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
 }
-//Notas para ti:
+//Notas:
 //1. categoryName: Se incluye para que el frontend pueda mostrar el nombre de la categoría sin tener que hacer una consulta adicional.
 //2. Sin validaciones: Al ser un DTO de respuesta, no lleva anotaciones de validación (@NotNull, etc.).
 //3. Tipos de datos: Asegúrate de que coincidan con la Entidad (BigDecimal para precio, Boolean para active).
